Add options to deal with repeated sentences
Now, there is the option to merge repeated sentences and get a single
output for them. You can choose to merge them while taking into account
tag proportions or not (i.e. if a token was marked once as B and four
times as O, you can decide if B get 0.5 and O gets 0.5 or if B gets 0.2
and O gets 0.8).