Add methods to generate a relationship triple CSV
These methods take a list of senteces, generate the relationship triples
and save them on the path passed as an argument.

The final step is to add methods to take a text file, split it into multiple
sentences and run them through this CSV generator, adding the title and
the unique ID prefix.