Fix the subset replacement
The replacement of subsets was being done by replacing the reference to
the subset list with another reference to the superset list. This
became a problem after the postprocessing was added in the form of
clean_triples and fix_partial_words. To fix this, it is necessary for
each list reference to appear only once, in its original spot, and for
its contents to be changed during the subset replacement.